renamed tableNames.ts
